Stage_Test/MagicLinkProject/terraform/modules/cloud_function/function_files/auth0_function/requirements.txt
flask==2.*
requests==2.*
python-jose==3.*
functions-framework==3.*
google-cloud-firestore==2.*